Handle files with spaces in their paths
Manipulation of arrays of paths broke if they contained spaces (or
tabulations). This is fixed by setting IFS (internal field separator) to
a single line break on array assignment.

This is done locally in 'array__from_line'. The propagation of the IFS
variable is prevented as it may cause unexpected consequences.